test(frontend): repoint stable's new tests at release-1.11's moved modules
The merge auto-merged cleanly but left two test files that stable added
importing modules release-1.11 had relocated, so the merged tree did not
compile. Repoint them:

- get-default-parent-from-api.test.ts: the Apollo client module
  `@/shared/api/graphql/graphqlClientApollo` (default export) became the named
  `graphqlClient` export of `@/shared/api/graphql/client`, and
  `@/entities/schema/domain/get-schema` moved under `domain/use-cases/`. The new
  client deliberately preserves the Apollo `query`/`mutate` surface, so both
  assertions still hold unchanged. `client` also re-exports gql.tada's `graphql`
  tag, which the module under test calls, so the mock factory keeps that export
  real and stubs only the transport — matching what the original test did.
- date-metadata-filter-form.test.tsx: `metadata-filter-definitions` moved from
  `domain/` to `domain/model/`.

Only import paths and the mock's export shape changed; no assertion was altered.
